requirements.txt
albumentations==1.4.2
anyio==4.4.0
astor==0.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
einops==0.2.0
filelock==3.15.4
fire==0.6.0
fonttools==4.53.1
fsspec==2024.6.1
gdown==4.4.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.2
imgaug==0.4.0
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.5.1
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.1
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle-gpu==2.6.1
pillow==10.2.0
prefetch_generator==1.0.1
protobuf==3.20.2
pyclipper==1.3.0.post5
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
PyYAML==6.0.1
rapidfuzz==3.9.4
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sympy==1.13.1
tbb==2021.13.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.7.21
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
vietocr==0.3.13
